+++ examples/shouldi/tests/test_npm_audit.py
@@ -1,31 +1,17 @@
 import pathlib
 
-from dffml import (
-    prepend_to_path,
-    cached_download_unpack_archive,
-    AsyncTestCase,
-)
+from dffml import prepend_to_path, AsyncTestCase
 
 from shouldi.javascript.npm_audit import run_npm_audit
 
+from .binaries import cached_node, cached_target_javascript_algorithms
+
 
 class TestRunNPMAuditOp(AsyncTestCase):
-    @cached_download_unpack_archive(
-        "https://nodejs.org/dist/v12.16.1/node-v12.16.1-linux-x64.tar.gz",
-        pathlib.Path(__file__).parent / "downloads" / "npm.tar.gz",
-        pathlib.Path(__file__).parent / "downloads" / "npm-audit-download",
-        "7df0e7b9f0d7e387c866c3b75596d924a63d11233e7a1a850acdeb333729ebbc9dcf01b1724ddf48a48bedf0cf2fddd8",
-    )
-    @cached_download_unpack_archive(
-        "https://github.com/trekhleb/javascript-algorithms/archive/ba2d8dc4a8e27659c1420fe52390cb7981df4a94.tar.gz",
-        pathlib.Path(__file__).parent / "downloads" / "javascript_algo.tar.gz",
-        pathlib.Path(__file__).parent
-        / "downloads"
-        / "javascript_algo-download",
-        "36b3ce51780ee6ea8dcec266c9d09e3a00198868ba1b041569950b82cf45884da0c47ec354dd8514022169849dfe8b7c",
-    )
-    async def test_run(self, npm_audit, javascript_algo):
-        with prepend_to_path(npm_audit / "bin"):
+    @cached_node
+    @cached_target_javascript_algorithms
+    async def test_run(self, node, javascript_algo):
+        with prepend_to_path(node / "node-v14.2.0-linux-x64" / "bin"):
             results = await run_npm_audit(
                 str(
                     javascript_algo