+++ model/tensorflow/tests/test_dnnc.py
@@ -4,6 +4,7 @@ import pathlib
 import tempfile
 from typing import Type
 
+from dffml import train, predict, accuracy
 from dffml.record import Record, RecordData
 from dffml.source.source import Sources
 from dffml.source.memory import MemorySource, MemorySourceConfig
@@ -83,25 +84,26 @@ class TestDNN(AsyncTestCase):
         self.assertEqual(config.classifications, [0, 1, 2])
         self.assertEqual(config.clstype, int)
 
-    async def test_00_train(self):
-        async with self.sources as sources, self.model as model:
-            async with sources() as sctx, model() as mctx:
-                await mctx.train(sctx)
-
-    async def test_01_accuracy(self):
-        async with self.sources as sources, self.model as model:
-            async with sources() as sctx, model() as mctx:
-                res = await mctx.accuracy(sctx)
-                self.assertGreater(res, 0.9)
-
-    async def test_02_predict(self):
-        a = Record("a", data={"features": {self.feature.name: 1}})
-        async with Sources(
-            MemorySource(MemorySourceConfig(records=[a]))
-        ) as sources, self.model as model:
-            target_name = model.config.predict.name
-            async with sources() as sctx, model() as mctx:
-                res = [record async for record in mctx.predict(sctx.records())]
-                self.assertEqual(len(res), 1)
+    async def test_model(self):
+        for i in range(0, 7):
+            await train(self.model, self.sources)
+            res = await accuracy(self.model, self.sources)
+            # Retry because of tensorflow intermitant low accuracy
+            if res <= 0.9 and i < 5:
+                print("Retry i:", i, "accuracy:", res)
+                self.model_dir.cleanup()
+                self.model_dir = tempfile.TemporaryDirectory()
+                self.model.config = self.model.config._replace(
+                    directory=self.model_dir.name
+                )
+                continue
+            self.assertGreater(res, 0.9)
+            a = Record("a", data={"features": {self.feature.name: 1}})
+            target_name = self.model.config.predict.name
+            res = [
+                record
+                async for record in predict(self.model, a, keep_record=True)
+            ]
+            self.assertEqual(len(res), 1)
             self.assertEqual(res[0].key, a.key)
             self.assertTrue(res[0].prediction(target_name).value)